+++ src/controllers/DriveController.java
@@ -31,12 +31,23 @@ public class DriveController {
 	public class Delete{
 		public String name;
 	}
+
+	private static boolean verifyUserRole(Request req, String forbiddenRole){
+		if(req.session(false).attribute("user_role").equals(forbiddenRole)){
+            return false;
+		}
+		return true;
+	}
 	public static Route getDrives = (Request request, Response response) -> {
 		response.type("application/json");
 		return App.g.toJson(App.driveService.getDrives(request.params("email")));
 	};
 	
 	public static Route addDrive = (Request request, Response response)->{
+		if(!verifyUserRole(request, "USER")){
+            response.status(403);
+            return "Forbidden";
+        }
 		response.type("application/json");
 		DriveToAdd drive = App.g.fromJson(request.body(), DriveToAdd.class);
 		String userEmail = request.session(false).attribute("email");
@@ -50,6 +61,10 @@ public class DriveController {
 		return App.g.toJson(null);
 	};
 	public static Route deleteDrive = (Request request, Response res)->{
+		if(!verifyUserRole(request, "USER")){
+            res.status(403);
+            return "Forbidden";
+        }
 		Delete name = App.g.fromJson(request.body(), Delete.class);
 		res.type("application/json");
 		if(App.driveService.deleteDrive(name.name)) {
@@ -64,6 +79,10 @@ public class DriveController {
 		return App.g.toJson(App.driveService.getAvilableDrives());
 	};
 	public static Route updateDrive = (Request req, Response res)->{
+		if(!verifyUserRole(req, "USER")){
+            res.status(403);
+            return "Forbidden";
+        }
 		res.type("application/json");
 		Update update = App.g.fromJson(req.body(), Update.class);
 		if(App.driveService.updateDrive(update)) {
